Move diff related types to ui-types
